[SCB-2094] Extract some common functions into mongo/util
