+++ datasource/mongo/dep.go
@@ -23,13 +23,11 @@ import (
 	"fmt"
 
 	"github.com/go-chassis/cari/discovery"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/mongo/options"
 
 	"github.com/apache/servicecomb-service-center/datasource"
-	"github.com/apache/servicecomb-service-center/datasource/etcd/path"
 	"github.com/apache/servicecomb-service-center/datasource/mongo/client"
 	"github.com/apache/servicecomb-service-center/datasource/mongo/db"
+	mutil "github.com/apache/servicecomb-service-center/datasource/mongo/util"
 	"github.com/apache/servicecomb-service-center/pkg/log"
 	"github.com/apache/servicecomb-service-center/pkg/util"
 )
@@ -37,8 +35,8 @@ import (
 func (ds *DataSource) SearchProviderDependency(ctx context.Context, request *discovery.GetDependenciesRequest) (*discovery.GetProDependenciesResponse, error) {
 	domainProject := util.ParseDomainProject(ctx)
 	providerServiceID := request.ServiceId
-	filter := GeneratorServiceFilter(ctx, providerServiceID)
-	provider, err := GetService(ctx, filter)
+	filter := mutil.GeneratorServiceFilter(ctx, providerServiceID)
+	provider, err := mutil.GetService(ctx, filter)
 	if err != nil {
 		if errors.Is(err, datasource.ErrNoData) {
 			log.Debug(fmt.Sprintf("query provider service failed, there is no provider %s in db", providerServiceID))
@@ -56,8 +54,8 @@ func (ds *DataSource) SearchProviderDependency(ctx context.Context, request *dis
 		}, nil
 	}
 
-	dr := NewProviderDependencyRelation(ctx, domainProject, provider.Service)
-	services, err := dr.GetDependencyConsumers(ToDependencyFilterOptions(request)...)
+	dr := mutil.NewProviderDependencyRelation(ctx, domainProject, provider.Service)
+	services, err := dr.GetDependencyConsumers(mutil.ToDependencyFilterOptions(request)...)
 	if err != nil {
 		log.Error(fmt.Sprintf("GetProviderDependencies failed, provider is %s/%s/%s/%s",
 			provider.Service.Environment, provider.Service.AppId, provider.Service.ServiceName, provider.Service.Version), err)
@@ -76,8 +74,8 @@ func (ds *DataSource) SearchConsumerDependency(ctx context.Context, request *dis
 	domainProject := util.ParseDomainProject(ctx)
 	consumerID := request.ServiceId
 
-	filter := GeneratorServiceFilter(ctx, consumerID)
-	consumer, err := GetService(ctx, filter)
+	filter := mutil.GeneratorServiceFilter(ctx, consumerID)
+	consumer, err := mutil.GetService(ctx, filter)
 	if err != nil {
 		if errors.Is(err, datasource.ErrNoData) {
 			log.Debug(fmt.Sprintf("query consumer service failed, there is no consumer %s in db", consumerID))
@@ -95,8 +93,8 @@ func (ds *DataSource) SearchConsumerDependency(ctx context.Context, request *dis
 		}, nil
 	}
 
-	dr := NewConsumerDependencyRelation(ctx, domainProject, consumer.Service)
-	services, err := dr.GetDependencyProviders(ToDependencyFilterOptions(request)...)
+	dr := mutil.NewConsumerDependencyRelation(ctx, domainProject, consumer.Service)
+	services, err := dr.GetDependencyProviders(mutil.ToDependencyFilterOptions(request)...)
 	if err != nil {
 		log.Error(fmt.Sprintf("query consumer failed, consumer is %s/%s/%s/%s",
 			consumer.Service.Environment, consumer.Service.AppId, consumer.Service.ServiceName, consumer.Service.Version), err)
@@ -129,7 +127,7 @@ func (ds *DataSource) AddOrUpdateDependencies(ctx context.Context, dependencys [
 			return rsp.Response, nil
 		}
 
-		consumerID, err := GetServiceID(ctx, consumerInfo)
+		consumerID, err := mutil.GetServiceID(ctx, consumerInfo)
 		if err != nil && !errors.Is(err, datasource.ErrNoData) {
 			log.Error(fmt.Sprintf("put request into dependency queue failed, override: %t, get consumer %s id failed",
 				override, consumerFlag), err)
@@ -161,7 +159,7 @@ func (ds *DataSource) AddOrUpdateDependencies(ctx context.Context, dependencys [
 			}
 			log.Info(fmt.Sprintf("insert dep to mongodb success %s", insertRes.InsertedID))
 		}
-		err = syncDependencyRule(ctx, domainProject, dependency)
+		err = mutil.SyncDependencyRule(ctx, domainProject, dependency)
 		if err != nil {
 			return nil, err
 		}
@@ -171,303 +169,4 @@ func (ds *DataSource) AddOrUpdateDependencies(ctx context.Context, dependencys [
 
 func (ds *DataSource) DeleteDependency() {
 	panic("implement me")
-}
-
-func syncDependencyRule(ctx context.Context, domainProject string, r *discovery.ConsumerDependency) error {
-
-	consumerInfo := discovery.DependenciesToKeys([]*discovery.MicroServiceKey{r.Consumer}, domainProject)[0]
-	providersInfo := discovery.DependenciesToKeys(r.Providers, domainProject)
-
-	var dep datasource.Dependency
-	//var err error
-	dep.DomainProject = domainProject
-	dep.Consumer = consumerInfo
-	dep.ProvidersRule = providersInfo
-	// add mongo get dep here
-
-	oldProviderRules, err := GetOldProviderRules(&dep)
-	if err != nil {
-		return err
-	}
-
-	if r.Override {
-		datasource.ParseOverrideRules(ctx, &dep, oldProviderRules)
-	} else {
-		datasource.ParseAddOrUpdateRules(ctx, &dep, oldProviderRules)
-	}
-	return updateDeps(domainProject, &dep)
-}
-
-func GetOldProviderRules(dep *datasource.Dependency) (*discovery.MicroServiceDependency, error) {
-	microServiceDependency := &discovery.MicroServiceDependency{
-		Dependency: []*discovery.MicroServiceKey{},
-	}
-	filter := GenerateConsumerDependencyRuleKey(dep.DomainProject, dep.Consumer)
-	findRes, err := client.GetMongoClient().FindOne(context.TODO(), db.CollectionDep, filter)
-	if err != nil {
-		log.Error(fmt.Sprintf("get dependency rule [%v] failed", filter), err)
-		return nil, err
-	}
-
-	if findRes.Err() != nil {
-		return microServiceDependency, nil
-	}
-
-	var depRule *db.DependencyRule
-	err = findRes.Decode(&depRule)
-	if err != nil {
-		return nil, err
-	}
-	return depRule.Dep, nil
-}
-
-func updateDeps(domainProject string, dep *datasource.Dependency) error {
-	var upsert = true
-	for _, r := range dep.DeleteDependencyRuleList {
-		filter := GenerateProviderDependencyRuleKey(domainProject, r)
-		_, err := client.GetMongoClient().Update(context.TODO(), db.CollectionDep, filter, bson.M{"$pull": bson.M{StringBuilder([]string{db.ColumnDep, db.ColumnDependency}): dep.Consumer}})
-		if err != nil {
-			return err
-		}
-		if r.ServiceName == "*" {
-			break
-		}
-	}
-	for _, r := range dep.CreateDependencyRuleList {
-		filter := GenerateProviderDependencyRuleKey(domainProject, r)
-		data := bson.M{
-			"$addToSet": bson.M{StringBuilder([]string{db.ColumnDep, db.ColumnDependency}): dep.Consumer},
-		}
-		_, err := client.GetMongoClient().Update(context.TODO(), db.CollectionDep, filter, data, &options.UpdateOptions{Upsert: &upsert})
-		if err != nil {
-			return err
-		}
-		if r.ServiceName == "*" {
-			break
-		}
-	}
-	filter := GenerateConsumerDependencyRuleKey(domainProject, dep.Consumer)
-	if len(dep.ProvidersRule) == 0 {
-		_, err := client.GetMongoClient().Delete(context.TODO(), db.CollectionDep, filter)
-		if err != nil {
-			return err
-		}
-	} else {
-		updateData := bson.M{
-			"$set": bson.M{StringBuilder([]string{db.ColumnDep, db.ColumnDependency}): dep.ProvidersRule},
-		}
-		_, err := client.GetMongoClient().Update(context.TODO(), db.CollectionDep, filter, updateData, &options.UpdateOptions{Upsert: &upsert})
-		if err != nil {
-			return err
-		}
-	}
-
-	err := CleanUpDepRules(context.TODO(), domainProject)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func CleanUpDepRules(ctx context.Context, domainProject string) error {
-	if len(domainProject) == 0 {
-		return ErrInvalidDomainProject
-	}
-
-	cache := make(map[*db.DelDepCacheKey]bool)
-	err := removeProviderRuleOfConsumer(ctx, domainProject, cache)
-
-	if err != nil {
-		return err
-	}
-
-	return removeProviderRuleKeys(ctx, domainProject, cache)
-}
-
-func removeProviderRuleOfConsumer(ctx context.Context, domainProject string, cache map[*db.DelDepCacheKey]bool) error {
-	filter := GenerateConsumerDependencyRuleKey(domainProject, nil)
-	depRules, err := GetDepRules(ctx, filter)
-	if err != nil {
-		return err
-	}
-	for _, depRule := range depRules {
-		err := removeConsumerDeps(ctx, depRule, cache)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func removeProviderRuleKeys(ctx context.Context, domainProject string, cache map[*db.DelDepCacheKey]bool) error {
-	filter := GenerateProviderDependencyRuleKey(domainProject, nil)
-	depRules, err := GetDepRules(ctx, filter)
-	if err != nil {
-		return err
-	}
-	for _, depRule := range depRules {
-		err := removeProviderDeps(ctx, depRule, cache)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func GetDepRules(ctx context.Context, filter bson.M) ([]*db.DependencyRule, error) {
-	findRes, err := client.GetMongoClient().Find(ctx, db.CollectionDep, filter)
-	if err != nil {
-		return nil, err
-	}
-
-	var depRules []*db.DependencyRule
-	for findRes.Next(ctx) {
-		var depRule *db.DependencyRule
-		err := findRes.Decode(&depRule)
-		if err != nil {
-			return nil, err
-		}
-		depRules = append(depRules, depRule)
-	}
-	return depRules, nil
-}
-
-func removeProviderDeps(ctx context.Context, depRule *db.DependencyRule, cache map[*db.DelDepCacheKey]bool) (err error) {
-	id := &db.DelDepCacheKey{
-		Key:  depRule.ServiceKey,
-		Type: path.DepsConsumer,
-	}
-	exist, ok := cache[id]
-	if !ok {
-		if depRule.ServiceKey.ServiceName == "*" {
-			return nil
-		}
-
-		_, exist, err = FindServiceIds(ctx, depRule.ServiceKey.Version, depRule.ServiceKey)
-		if err != nil {
-			return err
-		}
-
-		cache[id] = exist
-	}
-
-	filter := bson.M{
-		db.ColumnServiceKey: depRule.ServiceKey,
-	}
-	if !exist {
-		_, err = client.GetMongoClient().DocDelete(ctx, db.CollectionDep, filter)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func removeConsumerDeps(ctx context.Context, depRule *db.DependencyRule, cache map[*db.DelDepCacheKey]bool) (err error) {
-	var left []*discovery.MicroServiceKey
-	for _, key := range depRule.Dep.Dependency {
-		if key.ServiceName == "*" {
-			left = append(left, key)
-			continue
-		}
-
-		id := &db.DelDepCacheKey{
-			Key:  key,
-			Type: path.DepsProvider,
-		}
-		exist, ok := cache[id]
-		if !ok {
-			_, exist, err = FindServiceIds(ctx, key.Version, key)
-			if err != nil {
-				return err
-			}
-			cache[id] = exist
-		}
-
-		if exist {
-			left = append(left, key)
-		}
-	}
-	if len(depRule.Dep.Dependency) == len(left) {
-		return nil
-	}
-
-	filter := bson.M{
-		db.ColumnServiceKey: depRule.ServiceKey,
-	}
-	if len(left) == 0 {
-		_, err = client.GetMongoClient().DocDelete(ctx, db.CollectionDep, filter)
-	} else {
-		updateData := bson.M{
-			"$set": bson.M{StringBuilder([]string{db.ColumnDep, db.ColumnDependency}): left},
-		}
-		_, err = client.GetMongoClient().Update(ctx, db.CollectionDep, filter, updateData)
-	}
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func TransferToMicroServiceDependency(ctx context.Context, filter bson.M) (*discovery.MicroServiceDependency, error) {
-	microServiceDependency := &discovery.MicroServiceDependency{
-		Dependency: []*discovery.MicroServiceKey{},
-	}
-	findRes, err := client.GetMongoClient().FindOne(context.TODO(), db.CollectionDep, filter)
-	if err != nil {
-		return nil, err
-	}
-	if findRes.Err() == nil {
-		var depRule *db.DependencyRule
-		err := findRes.Decode(&depRule)
-		if err != nil {
-			return nil, err
-		}
-		microServiceDependency.Dependency = append(microServiceDependency.Dependency, depRule.Dep.Dependency...)
-		return microServiceDependency, nil
-	}
-	return microServiceDependency, nil
-}
-
-func GetConsumerDepInfo(ctx context.Context, filter bson.M) ([]*discovery.ConsumerDependency, error) {
-	var ConsumerDeps []*discovery.ConsumerDependency
-
-	findRes, err := client.GetMongoClient().Find(context.TODO(), db.CollectionDep, filter)
-	if err != nil {
-		return nil, err
-	}
-
-	for findRes.Next(ctx) {
-		var dep *db.ConsumerDep
-		err = findRes.Decode(&dep)
-		if err != nil {
-			return nil, err
-		}
-		ConsumerDeps = append(ConsumerDeps, dep.ConsumerDep)
-	}
-	return ConsumerDeps, nil
-}
-
-func GetServiceID(ctx context.Context, key *discovery.MicroServiceKey) (string, error) {
-	id, err := getServiceID(ctx, GeneratorServiceNameFilter(ctx, key))
-	if err != nil && !errors.Is(err, datasource.ErrNoData) {
-		return "", err
-	}
-	if len(id) == 0 && len(key.Alias) != 0 {
-		return getServiceID(ctx, GeneratorServiceAliasFilter(ctx, key))
-	}
-	return id, nil
-}
-
-func getServiceID(ctx context.Context, filter bson.M) (serviceID string, err error) {
-	svc, err := GetService(ctx, filter)
-	if err != nil {
-		return
-	}
-	if svc != nil {
-		serviceID = svc.Service.ServiceId
-		return
-	}
-	return
 }
